requirements.txt
discord.py==1.7.2
aiohttp==3.7.4
python-dateutil==2.8.1